requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication
firebase-admin==6.2.0
supabase==2.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Data processing
pandas==2.1.3
numpy==1.25.2
openpyxl==3.1.2

# Web scraping and HTTP
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10

# LLM integrations
openai==1.3.5
groq==0.4.1
google-generativeai==0.3.2

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS and middleware
fastapi-cors==0.0.6

# File handling
aiofiles==23.2.1

# Validation and utilities
email-validator==2.1.0
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.24.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

